requirements.txt
numpy==1.22.3
matplotlib==3.5.2
keras==2.9.0
scikit-learn==1.1.2
pyqt5==5.15.7
pillow==9.2.0
scipy==1.7.3
opencv-python==4.6.0.66
